requirements.txt
pyyaml
python-dotenv
openai
pygithub
apscheduler
